requirements.txt
groq==1.5.0
openai==1.12.0
fastapi==0.109.0
uvicorn==0.27.0
streamlit==1.30.0
python-dotenv==1.0.0
sqlalchemy==2.0.25
python-dateutil==2.8.2
APScheduler==3.10.4